requirements.txt
fastapi
uvicorn
google-generativeai
pillow
python-dotenv
python-multipart